fix: drive maintenance mode from DB for manager row id=2

+++ src/main/java/com/csee/swplus/mileage/setting/repository/SwManagerSettingRepository.java
@@ -4,6 +4,6 @@ import com.csee.swplus.mileage.setting.entity.SwManagerSetting;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface SwManagerSettingRepository extends JpaRepository<SwManagerSetting, Integer> {
+public interface SwManagerSettingRepository extends JpaRepository<SwManagerSetting, Long> {
     Optional<SwManagerSetting> findFirstByOrderByIdDesc();
 }
